runaba input restrict be float, get pcell, remove pdb

+++ ababe/cmdline/runabalib.py
@@ -58,7 +58,7 @@ def suplat(input, comment, volumn, zoom, ld, outmode):
 @click.option('--speckle', default=None)
 @click.option('--number-speckle', '-n', 'nspeckle', type=int, default=None)
 @click.option('--zoom', type=float, default=None)
-@click.option('--dist-restrict', '-r', 'trs', nargs=2, type=click.Tuple([str, int]), multiple=True)
+@click.option('--dist-restrict', '-r', 'trs', nargs=2, type=click.Tuple([str, float]), multiple=True)
 def ocumaker(input, comment, element, speckle, nspeckle, zoom, trs):
     infile = click.format_filename(input)
     y = yaml.load(open(infile, "r"))
